repo de chamados

+++ src/main/java/com/example/apiprojeto/repository/ChamadoRepository.java
@@ -3,5 +3,5 @@ package com.example.apiprojeto.repository;
 import com.example.apiprojeto.domain.Chamados;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChamadoRepository extends JpaRepository<Chamados, Integer> {
+public interface ChamadoRepository extends JpaRepository<Chamados, Long> {
 }
